@@ -1,7 +1,7 @@
 ﻿@using Valour.Shared
 @inject HttpClient Http
 @inject NavigationManager navManager
-@inject LocalStorageService storage
+@inject ILocalStorageService storage
 @inject AutoMapper.IMapper mapper
 @inject IJSRuntime JS
 
