@@ -2,7 +2,7 @@
 
 @inject HttpClient Http
 @inject NavigationManager navManager
-@inject LocalStorageService storage
+@inject ILocalStorageService storage
 @inject AutoMapper.IMapper Mapper
 @inject ClientWindowManager windowManager
 
