@@ -2,7 +2,7 @@
 
 @inject HttpClient Http
 @inject NavigationManager navManager
-@inject LocalStorageService storage
+@inject ILocalStorageService storage
 
 <body class="login-background" style="background-image: url(/media/Abstract-Background.png)">
     <div class="popup-box">
