@@ -2,7 +2,7 @@
 
 @inject HttpClient Http
 @inject NavigationManager navManager
-@inject LocalStorageService storage
+@inject ILocalStorageService storage
 @inject ClientPlanetManager planetManager
 @inject AutoMapper.IMapper mapper
 
