@@ -37,6 +37,14 @@ namespace Valour.Client.Planets
         /// </summary>
         private User _user = null;
 
+        public void RemoveRoleId(ulong RoleId)
+        {
+            if (_roleids != null)
+            {
+                _roleids.Remove(RoleId);
+            }
+        }
+
         public ulong TryGetPrimaryRoleId()
         {
             if (_roleids == null)
@@ -198,13 +206,13 @@ namespace Valour.Client.Planets
             _roleids = result.Data;
         }
 
-        public async Task<uint> GetAuthorityAsync()
+        public async Task<ulong> GetAuthorityAsync()
         {
             string json = await ClientUserManager.Http.GetStringAsync($"Planet/GetMemberAuthority?member_id={Id}&token={ClientUserManager.UserSecretToken}");
 
             Console.WriteLine($"Got authority for {Id}: " + json);
 
-            TaskResult<uint> result = JsonConvert.DeserializeObject<TaskResult<uint>>(json);
+            TaskResult<ulong> result = JsonConvert.DeserializeObject<TaskResult<ulong>>(json);
 
             if (result == null)
             {
