@@ -1,9 +1,9 @@
 @inject HttpClient Http
 @inject NavigationManager navManager
-@inject LocalStorageService storage
 @inject AutoMapper.IMapper Mapper 
 @inject ClientPlanetManager planetManager
 @inject IJSRuntime JS
+
 @using Valour.Shared.Categories
 
 <section>
