@@ -1,25 +1,59 @@
 @inject HttpClient Http
 @inject NavigationManager navManager
-@inject LocalStorageService storage
 @inject AutoMapper.IMapper Mapper 
 @inject ClientPlanetManager planetManager
 @inject AddChannelContextMenu contextMenu
 @inject IJSRuntime JS
+@inject CreateChannelModal createChannelModal
 @using Valour.Shared.Categories
 
 <section>
-    <h2>Create a new Channel</h2>
+    <h2 class="create-component-title">Create a Channel</h2>
 
-    <hr />
     <div asp-validation-summary="All" class="text-info"></div>
     <div class="form-group mt-2">
         <label>Channel Name</label>
         <input class="form-control" @bind-value="@name" />
-        <span id="name-span" style="color:#888888">@nameSpan</span>
     </div>
+    <div style="margin-top:20px;">
+        <div class="perm-list-name mb-1">
+            Private Channel
+        </div>
+        <div type="button" style="float: right">
+            <label class="switch">
+                <input type="checkbox" @onclick="PrivateSwitch" checked="@isprivate">
+                <span class="slider round"></span>
+            </label>
+        </div>
+    </div>
+
+    @if (isprivate) {
+        <label style="margin-top: 12px;">Who can see and send messages in this channel?</label>
+        <div class="create-component-roles-box">
+            @foreach (PlanetRole role in Roles)
+            {
+                <div class="role">
+                   <div class="perm-list-name mb-1" style="color: @role.GetColorHex();">
+                        @role.Name
+                    </div>
+                    <div type="button" style="float: right">
+                        <label class="switch">
+                            <input type="checkbox" OnStateChange="SwitchRole(role)" checked="@SelectedRoles[role]">
+                            <span class="slider round"></span>
+                        </label>
+                    </div> 
+                </div>
+
+            }
+        </div>
+    }
+
     <span id="image-span" style="color:#888888">@imageSpan</span>
+    
     <div class="form-group mt-4">
-        <button class="btn btn-primary mt-2" @onclick="OnClickSubmit">Create</button>
+        <center>
+            <button class="btn btn-primary mt-2" @onclick="OnClickSubmit">Create</button>
+        </center>
     </div>
 </section>
 @code {
@@ -27,8 +61,34 @@
     string name;
 
     // Spans
-    string nameSpan = "The name of the channel";
     string imageSpan = "";
+    Dictionary<PlanetRole, bool> SelectedRoles = new Dictionary<PlanetRole, bool>();
+
+    List<PlanetRole> Roles = new List<PlanetRole>();
+    bool isprivate = false;
+
+    protected override async Task OnInitializedAsync() {
+        Roles.Clear();
+        SelectedRoles.Clear();
+        ClientPlanetMember SelfMember = await ClientPlanetMember.GetClientPlanetMemberAsync(ClientUserManager.User.Id, planetManager.GetCurrent().Id);
+        ulong SelfAuthority = await SelfMember.GetAuthorityAsync();
+        foreach (PlanetRole item in await planetManager.GetCurrent().GetRolesAsync())
+        {
+            if (item.GetAuthority() >= SelfAuthority)
+                continue;
+            Roles.Add(item);
+            SelectedRoles.Add(item, false);
+        }
+        StateHasChanged();
+    }
+
+    private void SwitchRole(PlanetRole role) {
+        SelectedRoles[role] = !SelectedRoles[role];
+    }
+
+    private void PrivateSwitch() {
+        isprivate = !isprivate;
+    }
 
     private async Task OnClickSubmit(MouseEventArgs e)
     {
@@ -63,6 +123,69 @@
         }
 
         imageSpan = result.Message;
+
+        if (isprivate) {
+            // set the default role channel view perms to false
+
+            PermissionsNode PermissionsNode = new ChatChannelPermissionsNode()
+            {
+                Channel_Id = result.Data,
+                Planet_Id = planetManager.GetCurrent().Id,
+                Role_Id = planetManager.GetCurrent().Default_Role_Id,
+                Code_Mask = 0x00,
+                Code = 0x00
+            };
+
+            PermissionsNode.SetPermission(ChatChannelPermissions.PostMessages, PermissionState.False);
+            PermissionsNode.SetPermission(ChatChannelPermissions.ViewMessages, PermissionState.False);
+            PermissionsNode.SetPermission(ChatChannelPermissions.View, PermissionState.False);
+
+            HttpResponseMessage response = await ClientUserManager.Http.PostAsJsonAsync<ChatChannelPermissionsNode>($"Permissions/UpdateChatChannelNode?token={ClientUserManager.UserSecretToken}",
+                                                                                                    (ChatChannelPermissionsNode)PermissionsNode);
+
+            Console.WriteLine(result.Message);
+
+            foreach (PlanetRole role in Roles)
+            {
+                
+                if (!SelectedRoles[role]) {
+                    continue;
+                }
+                
+                PermissionsNode = new ChatChannelPermissionsNode()
+                {
+                    Channel_Id = result.Data,
+                    Planet_Id = planetManager.GetCurrent().Id,
+                    Role_Id = role.Id,
+                    Code_Mask = 0x00,
+                    Code = 0x00
+                };
+
+                PermissionsNode.SetPermission(ChatChannelPermissions.PostMessages, PermissionState.True);
+                PermissionsNode.SetPermission(ChatChannelPermissions.ViewMessages, PermissionState.True);
+                PermissionsNode.SetPermission(ChatChannelPermissions.View, PermissionState.True);
+
+                response = await ClientUserManager.Http.PostAsJsonAsync<ChatChannelPermissionsNode>($"Permissions/UpdateChatChannelNode?token={ClientUserManager.UserSecretToken}",
+                                                                                                        (ChatChannelPermissionsNode)PermissionsNode);
+                if (result == null)
+                {
+                    Console.WriteLine("Error deserializing response for permission node update!");
+                }
+
+                if (!result.Success)
+                {
+                    Console.WriteLine("Error in permission node update!");
+                }
+
+                Console.WriteLine(result.Message);
+            }
+        }
+
+        // close model
+
+        createChannelModal.Component.SetVisibility(false);
+
+        
     }
 
     private async Task OnImageInput()
