@@ -25,7 +25,10 @@
 </div>
 
 @code {
+ 
     public List<ClientPlanetMember> MemberList { get; set; }
+    public List<ClientPlanetMember> OrderedList { get; set; }
+
     public List<PlanetRole> PlanetRoles {get; set;}
 
     public Dictionary<PlanetRole, List<ClientPlanetMember>> RoleMemberLists { get; set; }
@@ -39,6 +42,48 @@
         OfflineList = new List<ClientPlanetMember>();
 
         planetManager.OnPlanetChange += OnPlanetChange;
+        planetManager.OnRoleUpdate += OnRoleUpdate;
+        planetManager.OnMemberUpdate += OnMemberUpdate;
+        planetManager.OnRoleDeletion += OnRoleDeletion;
+    }
+
+    public async Task OnMemberUpdate(ClientPlanetMember member)
+    {
+        // not awaited on purpose
+        if (member.Planet_Id == planetManager.GetCurrent().Id)
+        {
+            await RefreshOnRoleChange();
+        }
+        
+    }
+
+    public async Task OnRoleDeletion(PlanetRole role)
+    {
+        // not awaited on purpose
+        if (role.Planet_Id == planetManager.GetCurrent().Id)
+        {
+            await RefreshOnRoleChange();
+        }
+    }
+
+    public async Task OnRoleUpdate(PlanetRole role)
+    {
+        // not awaited on purpose
+        if (role.Planet_Id == planetManager.GetCurrent().Id)
+        {
+            await RefreshOnRoleChange();
+        }
+    }
+
+    public async Task RefreshOnRoleChange()
+    {
+        OrderedList = MemberList.OrderByDescending(GetComparator).ToList();
+        RoleMemberLists.Clear();
+        OnlineList.Clear();
+        OfflineList.Clear();
+        StateHasChanged();
+        await BuildRoleLists();
+        StateHasChanged();
     }
 
     public async Task BuildRoleLists()
@@ -47,7 +92,7 @@
         OnlineList.Clear();
         OfflineList.Clear();
 
-        foreach (var member in MemberList)
+        foreach (var member in OrderedList)
         {
             var user = await member.GetUserAsync();
 
@@ -101,9 +146,6 @@
 
     public async Task OnPlanetChange(ClientPlanet planet)
     {
-        // theres no reason to await the memberlist change
-        // not awaited on purpose
-
         Task.Run(async () => {
             Console.WriteLine($"Member list is updating to {planet.Name}");
             MemberList = await planetManager.GetCachedPlanetMembers(planet);
@@ -113,8 +155,6 @@
             Console.WriteLine($"Starting with {MemberList.Count()} members.");
             StateHasChanged();
         });
-
-        
     }
 
     public long GetComparator(ClientPlanetMember m)
