@@ -1,4 +1,4 @@
-﻿
+﻿@inject ClientPlanetManager planetManager
 @if (Role != null)
 {
 <p class="rolename" style="color: @Role.GetColorHex()">
