@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.SignalR;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,18 +14,13 @@ using System.Text.RegularExpressions;
 using System.Collections.Specialized;
 using Valour.Server.Messages;
 using Microsoft.EntityFrameworkCore;
-using System.Collections.Concurrent;
 using Valour.Server.Workers;
 using Valour.Server.Planets;
 using AutoMapper;
-using Valour.Server.MSP;
 using Valour.Server.Oauth;
 using Valour.Server.Categories;
 using WebPush;
-using Valour.Server.Notifications;
-using System.Web;
-using Microsoft.Extensions.Primitives;
-using System.Text.Json;
+using Valour.Server.MPS;
 
 /*  Valour - A free and secure chat client
  *  Copyright (C) 2021 Vooper Media LLC
@@ -99,12 +93,24 @@ namespace Valour.Server.Controllers
                 return new TaskResult(false, "You can not delete your main channel!");
             }
 
+            // remove permission nodes
+
+            Context.ChatChannelPermissionsNodes.RemoveRange(Context.ChatChannelPermissionsNodes.Where(x => x.Channel_Id == id));
+
+            await Context.SaveChangesAsync();
+
+            // remove all messages in this channel
+
+            Context.PlanetMessages.RemoveRange(Context.PlanetMessages.Where(x => x.Channel_Id == id));
+
+            await Context.SaveChangesAsync();
 
             Context.PlanetChatChannels.Remove(channel);
 
             await Context.SaveChangesAsync();
 
-            await PlanetHub.Current.Clients.Group($"p-{channel.Planet_Id}").SendAsync("RefreshChannelList", "");
+            // Notify of update
+            await PlanetHub.NotifyChatChannelDeletion(channel);
 
             return new TaskResult(true, "Successfully deleted.");
         }
@@ -194,7 +200,7 @@ namespace Valour.Server.Controllers
                 Parent_Id = parentid,
                 Message_Count = 0,
                 Description = "A chat channel",
-                Position = Convert.ToUInt16(Context.PlanetChatChannels.Count())
+                Position = (ushort)Context.PlanetChatChannels.Where(x => x.Parent_Id == parentid).Count()
             };
 
             // Add channel to database
@@ -413,7 +419,7 @@ namespace Valour.Server.Controllers
             
 
             // Media proxy layer
-            msg.Content = await MSPManager.HandleUrls(msg.Content);
+            msg.Content = await MPSManager.HandleUrls(msg.Content);
 
             PlanetMessageWorker.AddToQueue(msg);
 
