@@ -12,18 +12,17 @@ using Valour.Shared.Channels;
 using Valour.Shared.Categories;
 using Valour.Shared.Oauth;
 using Valour.Shared.Planets;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
-using Newtonsoft.Json;
 using System.Linq;
-using Valour.Server.MSP;
 using Valour.Server.Roles;
 using Valour.Shared.Roles;
 using Valour.Shared.Users;
 using Valour.Server.Users;
 using Valour.Server.Oauth;
 using Valour.Server.Categories;
+using Valour.Server.MPS.Proxy;
+using Valour.Server.MPS;
 
 
 /*  Valour - A free and secure chat client
@@ -151,15 +150,17 @@ namespace Valour.Server.Controllers
 
             // Use MSP for proxying image
 
-            MSPResponse proxyResponse = await MSPManager.GetProxy(image_url);
+            ProxyResponse proxyResponse = await MPSManager.GetProxy(image_url);
 
-            if (string.IsNullOrWhiteSpace(proxyResponse.Url) || !proxyResponse.Is_Media)
+            bool is_media = MPSManager.Media_Types.Contains(proxyResponse.Item.Mime_Type);
+
+            if (proxyResponse.Item == null || !is_media)
             {
                 image_url = "https://valour.gg/image.png";
             }
             else
             {
-                image_url = proxyResponse.Url;
+                image_url = proxyResponse.Item.Url;
             }
 
             ulong planet_id = IdManager.Generate();
@@ -1049,13 +1050,13 @@ namespace Valour.Server.Controllers
         /// <summary>
         /// Returns the authority of the requested member
         /// </summary>
-        public async Task<TaskResult<uint>> GetMemberAuthority(ulong member_id, string token)
+        public async Task<TaskResult<ulong>> GetMemberAuthority(ulong member_id, string token)
         {
             ServerAuthToken authToken = await ServerAuthToken.TryAuthorize(token, Context);
 
             if (authToken == null)
             {
-                return new TaskResult<uint>(false, "Failed to authorize user.", 0);
+                return new TaskResult<ulong>(false, "Failed to authorize user.", 0);
             }
 
             ServerPlanetMember member = await Context.PlanetMembers.Include(x => x.Planet)
@@ -1063,15 +1064,15 @@ namespace Valour.Server.Controllers
 
             if (member == null)
             {
-                return new TaskResult<uint>(false, "Member does not exist.", 0);
+                return new TaskResult<ulong>(false, "Member does not exist.", 0);
             }
 
             if (!(await member.Planet.IsMemberAsync(authToken.User_Id, Context)))
             {
-                return new TaskResult<uint>(false, "You are not in the planet.", 0);
+                return new TaskResult<ulong>(false, "You are not in the planet.", 0);
             }
 
-            return new TaskResult<uint>(true, "Found authority", await member.GetAuthorityAsync());
+            return new TaskResult<ulong>(true, "Found authority", await member.GetAuthorityAsync());
         }
 
         public async Task<TaskResult> InsertCategory(ulong category_id, ulong planet_id, ushort position, string auth)
