@@ -4,18 +4,21 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Valour.Server.MSP
+namespace Valour.Server.MPS
 {
-    public class MSPConfig
+    public class MPSConfig
     {
-        public static MSPConfig Current;
+        public static MPSConfig Current;
 
-        public MSPConfig()
+        public MPSConfig()
         {
             Current = this;
         }
 
         [JsonProperty]
         public string Api_Key { get; set; }
+
+        [JsonIgnore]
+        public string Api_Key_Encoded { get; set; }
     }
 }
